minimal demo, retain prevHistory
