+++ deno/v0.1/mod.ts
@@ -1,9 +1,6 @@
 export type { BroadcasterInterface, BroadcastMessage } from "./type_flyweight/router.ts";
 
 export {
-  BROADCAST,
-  HASH_CHANGE,
   push,
   setBroadcaster,
-  UNKNOWN,
 } from "./router/router.ts";
