+++ deno/v0.1/type_flyweight/router.ts
@@ -1,0 +1,12 @@
+type BroadcastMessage<D = unknown> = {
+  location: string;
+  title: string;
+  data?: D;
+};
+
+interface BroadcasterInterface<D = unknown> {
+  postMessage(message: BroadcastMessage<D>): void;
+}
+
+
+export type { BroadcasterInterface, BroadcastMessage }
