--- docs/deno/faux_data.ts
@@ -1,15 +1,0 @@
-const urlTitles: Record<string, string> = {
-  "/": "root",
-  "/#/dogs": "dogs",
-  "/#/cats": "cats",
-  "/#/pigs": "pigs",
-};
-
-const urlData: Record<string, string> = {
-  "/": "beasts tread softly underfoot",
-  "/#/dogs": "dogs crave companions and magic",
-  "/#/cats": "cats become grand hunters, do not lie to them",
-  "/#/pigs": "pigs are curious, often playful",
-};
-
-export { urlData, urlTitles };
