+++ docs/index.html
@@ -10,15 +10,10 @@
 </head>
 
 <body>
-    <div class="frame">
-        <!-- comment out elements below to run on localhost -->
-        <demo-menu path="/router/"></demo-menu>
-        <demo-display path="/router/"></demo-display>
-
-        <!-- uncomment elements below to run on localhost -->
-        <!-- <demo-menu></demo-menu>
-        <demo-display></demo-display> -->
-    </div>
+		<button>add random history</button>
+		<section>
+			<p>history state object will reside here!</p>
+		</section>
 </body>
 
 </html>
